moved codes to components and data
